csharp/interest-is-interesting/InterestIsInteresting/SavingsAccount.cs
using System;
namespace InterestIsInteresting
{
    public static class SavingsAccount
    {
        public static float InterestRate(decimal balance)
        {
            if (balance < 0) return 3.213f;
            else if (balance < 1000) return 0.5f;
            else if (balance >= 1000 && balance < 5000) return 1.621f;
            else return 2.475f;
        }

        public static decimal Interest(decimal balance)
        {
            throw new NotImplementedException("Please implement the (static) SavingsAccount.Interest() method");
        }

        public static decimal AnnualBalanceUpdate(decimal balance)
        {
            throw new NotImplementedException("Please implement the (static) SavingsAccount.AnnualBalanceUpdate() method");
        }

        public static int YearsBeforeDesiredBalance(decimal balance, decimal targetBalance)
        {
            throw new NotImplementedException("Please implement the (static) SavingsAccount.YearsBeforeDesiredBalance() method");
        }
    }
}
